Fixed game saving/loading, added Blueprints, more
Game saves and loads properly. World generation stores each region in a separate file to be loaded as the player explores. Towns generate properly. Blueprints were added that allow multiple entities to be placed as a set. Added a starter farm with two sheep. Continued work on blocks and logic handling. Fixed terrain_corners.png

+++ src/main/java/com/bitbucket/computerology/world/terrain/generators/Generator.java
@@ -21,13 +21,11 @@ public class Generator {
     /**
      * Applies a circular brush to the terrain.
      *
-     * @param os_x     Center of the terrain circle. On-screen coords.
-     * @param os_y     Center of the terrain circle.
      * @param diametre The diametre of the circle.
      * @param terrain  Terrain type to paint.
      * @param biome    Do you want to paint the chunk's biome or the terrain?
      */
-    public static void brush(int wx, int wy, int diametre, int terrain, boolean biome) {
+    public static void brush(int wx, int wy, int diametre, byte terrain, boolean biome) {
         for (int w = (-diametre / 2); w != (diametre / 2); w++) {
             for (int h = (-diametre / 2); h != (diametre / 2); h++) {
                 if (MiscMath.distance(w, h, 0, 0) <= (diametre / 2)) {
